apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.12.3,<3.0.0
pydantic-settings>=2.3.4
python-multipart==0.0.9
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
playwright>=1.40.0
cachetools>=5.3.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langsmith>=0.1.0
langgraph-checkpoint-postgres>=2.0.0  # For Postgres persistence

# Database
psycopg2-binary>=2.9.0  # Postgres driver (for thread_metadata, arena_service)
psycopg>=3.1.0  # Postgres driver for PostgresSaver (langgraph-checkpoint-postgres)

# EXA for web search
exa-py>=1.0.0

# Export utilities
weasyprint>=60.0  # For PDF export
markdown>=3.5.0

# Authentication
pyjwt>=2.8.0  # JWT token handling
resend>=0.8.0  # Email sending (optional)
email-validator>=2.1.0  # Email validation

# Testing
pytest-asyncio>=1.0.0  # Async test support
